chore(deps): break SyncEngine ↔ conflictResolution import cycle
Extract the four storage-agnostic callback types (ContentHashFn,
ContentDataFn, ItemNameFn, ContentApplyFn) into a leaf
SyncEngine.types.ts module. SyncEngine.conflictResolution now imports
ContentApplyFn from the leaf instead of from SyncEngine, breaking the
cycle. SyncEngine.ts re-exports all four names so external consumers
(SyncEngine tests, the vault barrel index, vault store wiring) keep
their existing import paths working.

Cycle removed: vault/SyncEngine.conflictResolution.ts ↔ vault/SyncEngine.ts
(madge --circular: 10 → 9)

No behavior change.
