+++ src/services/vault/SyncEngine.ts
@@ -15,6 +15,13 @@ import type {
   ChangeType,
 } from '@/types/vault';
 
+import type {
+  ContentApplyFn,
+  ContentDataFn,
+  ContentHashFn,
+  ItemNameFn,
+} from './SyncEngine.types';
+
 import {
   createSingleton,
   type SingletonFactory,
@@ -61,41 +68,10 @@ export interface IReconciliationResult {
   skipped: IChangeLogEntry[];
 }
 
-/**
- * Content hash function type
- */
-export type ContentHashFn = (
-  itemId: string,
-  contentType: ShareableContentType | 'folder',
-) => Promise<string | null>;
-
-/**
- * Content data function type
- */
-export type ContentDataFn = (
-  itemId: string,
-  contentType: ShareableContentType | 'folder',
-) => Promise<string | null>;
-
-/**
- * Item name resolver — returns a human-readable display name for a
- * conflicting item so conflict records aren't labelled with raw ids.
- */
-export type ItemNameFn = (
-  itemId: string,
-  contentType: ShareableContentType | 'folder',
-) => Promise<string | null>;
-
-/**
- * Content apply function — writes remote content back to local storage.
- * Registered by the caller that owns persistence (e.g. a vault store)
- * so the sync engine stays storage-agnostic.
- */
-export type ContentApplyFn = (
-  itemId: string,
-  contentType: ShareableContentType | 'folder',
-  data: string,
-) => Promise<void>;
+// Re-exported from SyncEngine.types so existing consumers that import these
+// callback types from `@/services/vault/SyncEngine` (or via the vault barrel
+// index) keep working unchanged.
+export type { ContentHashFn, ContentDataFn, ItemNameFn, ContentApplyFn };
 
 // =============================================================================
 // Service
